Make Job id work with tests
Modify how Job is assigned.  We had a problem if tests were run
in different order, they would get a different Job id, but we
were using this in tests of the string output of the queues.  We need
to be able to predict the id of a Job item, so we changed constructor
to explicitly need the id to be specified for new Job items.  This
makes it possible to test queues of Job items and know what id we
will expect for each Job, even if we run tests in different
orders.